Now MDPFunction exists.
Basically the rub is that the value function and policy function for a given MDP is really a dict under the hood. But conceptually these are functions, i.e. callable objects, and it's bad form to think of them as callable objects but implement them specifically with __getitem__ instead of __call__. It was causing problems, and now it's fixed.